Fix some performance issues with gdiff
We memoise cost instead of recalculating every time.

However, this uncovers that most of the time we're
spending marshalling around proof objects to satisfy the IsList
constraint. However, do we even need this constraint in modern
haskell, or can we hack our way around it? that's the biggest
question I have right now.  As now 50% of the time is spent
garbage collecting proof objects.