WIP: Further simplify gdiff, and prepare for CpyTree
The Cof and Tyof stuff was messy,  the aptly named "Lol" (WIP name)
is a better Cof that carries exactly the proofs we need,
such that we do not need the Tyof type family.

Furthermore means   there exists a function now
that does  NA -> Lol , which means we can simplify diffT

Furthermore, I've found a way to add CpyTree to edit script
and still make 'em typecheck .This will make room for the
Merkle tree optimisation that I had planned quite a long time ago
and might give us a significant performance boost and might
compact some memory as well!

In hindsight I was totally mindblocked on this 6 months ago, whilst
the answer was literally trivial. Not looking at your code for
6 months and then looking back really does wonders (Who the fuck wrote
this code??? Oh wait, it was me, as i'm the only one writing on this
project)