Further reimplement gdiff
The route to 'Lol' was the wrong one, as i'm not carrying the
right proof info. Cof is here to stay. it looks the most
like the Agda code.

However, implementing gdiff with a lot less proof objects
still seems to be worthwhile. We accomplished this
by letting Cof carry the list lenghth proof of its arguments